#74 DictionaryDisposable should have method containsKey(key)
#73 top level export should export DisposableBase
