--- src/model/ModelBase.js
@@ -1,36 +1,0 @@
-// notice_start
-/*
- * Copyright 2015 Dev Shop Limited
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
- // notice_end
-
-/**
- * A base class for model entities.
- *
- * You don't need to derive from this to use the router, provided as a convenience
- */
-export default class ModelBase {
-    constructor() {
-        this._checkIsLocked = () => true;
-    }
-    ensureLocked() {
-        if(this._checkIsLocked()) {
-            throw new Error("Model is locked, can't edit");
-        }
-    }
-    get isLocked() {
-        return this._checkIsLocked();
-    }
-}
