+++ src/system/disposables/index.js
@@ -19,3 +19,4 @@
 export { default as CompositeDisposable } from './CompositeDisposable';
 export { default as DictionaryDisposable } from './DictionaryDisposable';
 export { default as DisposableWrapper } from './DisposableWrapper';
+export { default as DisposableBase } from './DisposableBase';
